Add RPS model in script

+++ scripts/run_preprocessing.py
@@ -18,6 +18,9 @@ if narg > 1:
         base='/u/ckim14/'
     elif system =='tigress':
         base='/tigress/changgoo/'
+    elif system =='tigress_arm':
+        base='/tigress/changgoo/ARM/'
+        vtkbase='/tigress/woongkim/TWO/'
     elif system =='perseus':
         base='/perseus/scratch/gpfs/changgoo/'
     elif system =='cori':
@@ -51,43 +54,47 @@ for problem_id in ids:
     if (system == 'cori') | (system == 'rusty'):
         preprocessing.doall(base,problem_id,problem_dir='{}/slab/'.format(problem_id),\
                             do_pickling=do_pickling,use_yt=False)
+    elif (system == 'tigress_arm'):
+        preprocessing.doall(base,problem_id,do_pickling=do_pickling,use_yt=False,vtkdir=vtkbase,force_recal=True)
     else:
-        preprocessing.doall(base,problem_id,do_pickling=do_pickling,use_yt=False)
+        preprocessing.doall(base,problem_id,do_pickling=do_pickling,use_yt=False,force_recal=True)
 
 for problem_id in ids:
     if (system == 'cori') | (system == 'rusty'):
         problem_dir='{}/slab/'.format(problem_id)
     else:
         problem_dir='{}/'.format(problem_id)
-    print 'drawing {} ...'.format(problem_id)
-    h_zp=pd.read_pickle('{}{}hst/{}.hst_zp.p'.format(base,problem_dir,problem_id))
-    sfrmean=h_zp['sfr10'].mean()
-    snrmean=h_zp['snr10'].mean()
-    Pmidmean=h_zp['Pmid_2p'].mean()
-    metadata=[
-            ('surf',labels['surf']+label_units['surf'],'linear',[],None,None),
-            ('sfr10',labels['sfr']+label_units['sfr'],'log',['sfr40','sfr100'],sfrmean*0.05,None),
-            ('snr10',labels['snr']+label_units['snr'],'log',['snr40','snr100'],snrmean*0.05,None),
-            ('H',labels['scaleH']+label_units['scaleH'],'linear',['H_2p','H_c','H_u','H_w'],None,None),
-            ('sigma_eff_2p',labels['sigma_z']+label_units['velocity'],'linear',['v3_2p','vA_2p','cs_2p'],None,None),
-            ('Pmid_2p',labels['pressure_mid_2p']+label_units['pressure'],'log',
-             ['Pturb_mid_2p','Pth_mid_2p','Pimag_mean_mid_2p','Pimag_turb_mid_2p'],Pmidmean*0.01,None),
-            ('Pmid_2p',labels['pressure']+label_units['pressure'],'linear',['PDE2','W_2p'],None,None),
-            ('massflux_bd_d',labels['massflux']+label_units['massflux'],'linear',
-             ['massflux_bd_d_h','massflux_bd_d_2p'],None,None),
-            ('snr10',labels['metalflux']+label_units['massflux'],'linear',
-             ['massflux_bd_s1','massflux_bd_s2','massflux_bd_s3'],None,None),
-            ('sfr10',labels['massflux']+label_units['massflux'],'linear',
-             ['massflux_bd_d_h','massflux_out_5_h','massflux_out_10_h'],None,None),
-            ('mf_c',labels['massfrac'],'linear',
-             ['mf_u','mf_w'],None,None),
-             ]
-    figdir='{}{}/figures/'.format(base,problem_dir)
-    if not os.path.isdir(figdir): os.mkdir(figdir)
-    figfname='{}{}/figures/{}-history.png'.format(base,problem_dir,problem_id)
-    preprocessing.draw_history(h_zp,metadata,figfname)
-    if system == 'tigress':
-        figdir_tigress='{}/public_html/TIGRESS_figures/history/'.format(base)
-        if len(figfname) > 0: 
-            os.chmod(figfname,0644)
-            shutil.copy2(figfname,figdir_tigress)
+    hzpfile='{}{}hst/{}.hst_zp.p'.format(base,problem_dir,problem_id)
+    if os.path.isfile(hzpfile):
+        print 'drawing {} ...'.format(problem_id)
+        h_zp=pd.read_pickle('{}{}hst/{}.hst_zp.p'.format(base,problem_dir,problem_id))
+        sfrmean=h_zp['sfr10'].mean()
+        snrmean=h_zp['snr10'].mean()
+        Pmidmean=h_zp['Pmid_2p'].mean()
+        metadata=[
+                ('surf',labels['surf']+label_units['surf'],'linear',[],None,None),
+                ('sfr10',labels['sfr']+label_units['sfr'],'log',['sfr40','sfr100'],sfrmean*0.05,None),
+                ('snr10',labels['snr']+label_units['snr'],'log',['snr40','snr100'],snrmean*0.05,None),
+                ('H',labels['scaleH']+label_units['scaleH'],'linear',['H_2p','H_c','H_u','H_w'],None,None),
+                ('sigma_eff_2p',labels['sigma_z']+label_units['velocity'],'linear',['v3_2p','vA_2p','cs_2p'],None,None),
+                ('Pmid_2p',labels['pressure_mid_2p']+label_units['pressure'],'log',
+                 ['Pturb_mid_2p','Pth_mid_2p','Pimag_mean_mid_2p','Pimag_turb_mid_2p'],Pmidmean*0.01,None),
+                ('Pmid_2p',labels['pressure']+label_units['pressure'],'linear',['PDE2','W_2p'],None,None),
+                ('massflux_bd_d',labels['massflux']+label_units['massflux'],'linear',
+                 ['massflux_bd_d_h','massflux_bd_d_2p'],None,None),
+                ('snr10',labels['metalflux']+label_units['massflux'],'linear',
+                 ['massflux_bd_s1','massflux_bd_s2','massflux_bd_s3'],None,None),
+                ('sfr10',labels['massflux']+label_units['massflux'],'linear',
+                 ['massflux_bd_d_h','massflux_out_5_h','massflux_out_10_h'],None,None),
+                ('mf_c',labels['massfrac'],'linear',
+                 ['mf_u','mf_w'],None,None),
+                 ]
+        figdir='{}{}/figures/'.format(base,problem_dir)
+        if not os.path.isdir(figdir): os.mkdir(figdir)
+        figfname='{}{}/figures/{}-history.png'.format(base,problem_dir,problem_id)
+        preprocessing.draw_history(h_zp,metadata,figfname)
+        if system == 'tigress':
+            figdir_tigress='{}/public_html/TIGRESS_figures/history/'.format(base)
+            if len(figfname) > 0: 
+                os.chmod(figfname,0644)
+                shutil.copy2(figfname,figdir_tigress)
